feat: add command for main.py script
